requirements.txt
# Core dependencies
psycopg2-binary>=2.9.0

# Optional LLM support
google-generativeai>=0.3.0
anthropic>=0.7.0

# Optional formatting
tabulate>=0.9.0
